+++ Assets/Scripts/Enemies/TurretEnemy.cs
@@ -2,6 +2,6 @@
 
 public class TurretEnemy : Enemy
 {
-    public TurretEnemy(float hp, float rotationSpeed) : base(hp, rotationSpeed) { }
+    public TurretEnemy(float hp, float rotationSpeed) : base(hp) { }
     
 }
